requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
redis
django-redis
drf-yasg
requests
python-dotenv>=1.0
django-filter